// src/app/services/openapi-client/model/employeeAuthResponseDto.ts
/**
 * ShopErpApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EmployeeAuthDto } from './employeeAuthDto';


export interface EmployeeAuthResponseDto { 
    token?: string | null;
    refreshToken?: string | null;
    expiresAt?: string;
    employee?: EmployeeAuthDto;
}