// src/app/services/openapi-client/model/refreshTokenRequestDto.ts
/**
 * ShopErpApi | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * DTO for refresh token request
 */
export interface RefreshTokenRequestDto { 
    refreshToken: string;
}